[#2] refactor: core 패키지 삭제, 패키지명 변경
