+++ core/src/main/kotlin/site/snaplog/entity/BaseEntity.kt
@@ -1,4 +1,4 @@
-package site.snaplog.entities
+package site.snaplog.entity
 
 import org.springframework.data.annotation.CreatedDate
 import org.springframework.data.annotation.LastModifiedDate
