+++ core/src/main/kotlin/site/snaplog/entity/MemberEntity.kt
@@ -1,4 +1,4 @@
-package site.snaplog.entities
+package site.snaplog.entity
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Table
