+++ core/src/main/kotlin/site/snaplog/entity/SubjectEntity.kt
@@ -1,4 +1,4 @@
-package site.snaplog.entities
+package site.snaplog.entity
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Column
